backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
pydantic
motor
python-dotenv
